Stdlib. getchar() returns 0..0xFF

+++ include/arch/cpm/stdio_h/getchar.c
@@ -22,6 +22,6 @@ int getchar(void) {
 #ifdef __C8080_USE_CPM_CONSOLE_IO /* Cannot use CpmBiosConSt() because symbols get stuck in CP/M */
     return (uint8_t)CpmConsoleDirect(0xFF);
 #else
-    return CpmBiosConIn();
+    return (uint8_t)CpmBiosConIn();
 #endif
 }
